ad 102 fabionacciu fix1

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,14 +11,14 @@ int main(void)
 	int i;
 	int a = 0;
 	int b = 1;
-	unsigned int res = 0;
+	long unsigned int res = 0;
 
 	for (i = 1; i <= 50; i++)
 	{
 		res = a + b;
 		a = b;
 		b = res;
-		printf("%u, ", res);
+		printf("%lu, ", res);
 
 	}
 	printf("\n");
